+++ src/main/java/calculator/Calculator.java
@@ -4,6 +4,6 @@ import java.util.List;
 
 public interface Calculator {
 
-    long sum(List<Integer> numbers);
+    int sum(List<Integer> numbers);
 
 }
