+++ src/main/java/calculator/OutputView.java
@@ -2,7 +2,7 @@ package calculator;
 
 public class OutputView {
 
-    public void getSumResult(long sum) {
+    public void getSumResult(int sum) {
         System.out.println("결과 : " + sum);
     }
 }
